requirements.txt
apache-airflow==2.10.3
apache-airflow-providers-postgres==5.6.0
python-dotenv==1.0.1
kaggle==1.6.10
scikit-learn==1.3.2
pendulum==2.1.2
pandas==1.5.3
sqlalchemy==1.4.49
psycopg2-binary==2.9.9